Added singlitico and basic cost model files
